erp_by/backend/requirements.txt
flask
pyjwt
flask-cors
werkzeug
pymongo
python-dotenv
